# hoodscan/apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hoodscan API
  version: 1.0.0
  description: >
    Read API for the Hoodscan block explorer (Robinhood Chain, an Arbitrum-style
    L2). All list endpoints share a `{ ..., total, limit, offset }` envelope,
    large integers (wei / token amounts / block numbers) are serialized as
    strings, addresses are lowercase, and errors use `{ "error": string }`.
    A global rate limit of 120 requests/minute per IP applies to all endpoints
    except `/health`, `/docs` and `/openapi.json`.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /tokens/{address}/logo:
    get:
      tags: [Tokens]
      summary: Token logo image (proxied)
      parameters:
        - $ref: "#/components/parameters/AddressPath"
      responses:
        "200":
          description: Binary image of the token logo
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
...
    AddressPath:
      name: address
      in: path
      required: true
      description: 20-byte hex address (0x-prefixed). Lowercased server-side.
      schema: { type: string, example: "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2" }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Not found" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...